fixed jediparrot and sithparrot (#563)
* Delete jediparrot.gif

When uploaded to Discord, the gif unintentionally leaves a trail of images. Will upload a fixed gif.

* Delete sithparrot.gif

When uploaded to Discord, the gif unintentionally leaves a trail of images. Will upload a fixed gif.

* add fixed jediparrot and sithparrot

Previous gifs would leave a trail of images when uploaded to Discord.  The issue is fixed with these gifs.